Move furry detection to separate file
